Integrate: [ 27962] Subject: Re: [PATCH] cleanup 212 warnings emitted by gcc-4.2 Message-ID: <20060424232038.7550f9b6@r2d2> p4raw-link: @27962 on //depot/perl: bb7a0f5422600fd7de7806825a1300a577b7e61f
p4raw-id: //depot/maint-5.8/perl@29962
p4raw-integrated: from //depot/perl@27962 'edit in' op.c regexec.c
	(@27958..) 'merge in' patchlevel.h (@27007..) deb.c (@27300..)
	handy.h (@27774..) scope.h (@27805..) pp_sys.c (@27837..)
	util.c (@27908..) dump.c (@27910..) perl.c (@27941..) doop.c
	(@27943..) mg.c (@27946..) doio.c (@27951..) gv.c perlio.c pp.c
	pp_hot.c regcomp.c sv.c toke.c (@27958..) pp_ctl.c (@27959..)

+++ regcomp.c
@@ -4777,7 +4777,7 @@ Perl_regprop(pTHX_ SV *sv, regnode *o)
 	}
 
 	if (o->flags & ANYOF_CLASS)
-	    for (i = 0; i < sizeof(anyofs)/sizeof(char*); i++)
+	    for (i = 0; i < (int)(sizeof(anyofs)/sizeof(char*)); i++)
 		if (ANYOF_CLASS_TEST(o,i))
 		    sv_catpv(sv, anyofs[i]);
